(v0.0.1) FEAT: Adding VideoMoment - Cleaning up code organization to improve importing
