+++ .gitignore
@@ -18,6 +18,7 @@ dist
 .env.development.local
 .env.test.local
 .env.production.local
+report.*
 
 npm-debug.log*
 yarn-debug.log*
