+++ src/Components/Moment/Base/style.scss
@@ -1,0 +1,5 @@
+@import "../style.scss";
+
+.story-moment-base {
+  text-align: center;
+}
