+++ src/Components/Moment/Video/Readme.md
@@ -1,0 +1,5 @@
+`VideoMoment` Example:
+
+```js
+<VideoMoment title="The Original Video Game 🕹️"  subtitle="Embedding a video directly into a moment" url="https://upload.wikimedia.org/wikipedia/commons/7/7b/The_Original_Video_Game.webm" color={{background:'#d7f0fa', text:'#525252'}} />
+```
