+++ src/Components/Moment/Video/index.js
@@ -5,9 +5,9 @@ import MomentBase from '../Base';
 import './style.scss';
 
 /**
- * General iFrame layout Moment.
+ * General Video layout Moment.
  */
-export default class IFrameMoment extends Component {
+export default class VideoMoment extends Component {
   static propTypes = {
     /** Determines the background and text color of the `Moment` header. */
     color: PropTypes.shape({
@@ -22,25 +22,23 @@ export default class IFrameMoment extends Component {
     title: PropTypes.node,
     /** The type of `Moment` */
     type: PropTypes.string,
-    /** The source of the iFrame to render in the `Moment`'s body. */
+    /** The source of the Video to render in the `Moment`'s body. */
     url: PropTypes.string.isRequired,
 
   }
   static defaultProps = {
-    color: {
-      background: "gray",
-      text: "white"
-    },
-    type: "iframe",
+    type: "video",
   }
 
   render() {
-    const layout="iframe"; // TODO: Add constant
+    const layout="video"; // TODO: Add constant
     const { url } = this.props;
     return (
       <MomentBase {...this.props} layout={layout}>
-        <iframe src={url} width="100%" height="100vh" allowFullScreen
-          frameBorder="0"></iframe>
+        <video width="100%" controls>
+            <source src={url} />
+          Your browser does not support the video tag.
+        </video>
       </MomentBase>
     )
   }
