+++ src/Components/Moment/style.scss
@@ -27,7 +27,3 @@
     }
   }
 }
-
-.story-moment-base {
-  text-align: center;
-}
