+++ src/Components/index.js
@@ -1,0 +1,2 @@
+import * as Moments from './Moment';
+export { Moments };
