+++ src/index.js
@@ -1,3 +1,4 @@
 import './theme.scss';
-export MomentBase from './Components/Moment';
-export WikidataMoment from './Components/Moment/Wikidata'
+
+import * as Components from './Components';
+export { Components };
